+++ tests/testthat/test-pkg-MASS.R
@@ -118,7 +118,6 @@ test_that("marginaleffects vs. emmeans", {
 ### predictions
 
 test_that("polr: predictions: no validity", {
-    skip_if_not_installed("insight", minimum_version = "0.14.4.1")
     mod <- MASS::polr(factor(gear) ~ mpg + factor(cyl), data = mtcars)
     pred <- predictions(mod, type = "probs")
     expect_predictions(pred)
@@ -133,7 +132,6 @@ test_that("glm.nb: predictions: no validity", {
 })
 
 test_that("rlm: predictions: no validity", {
-    skip_if_not_installed("insight", minimum_version = "0.14.4.1")
     model <- MASS::rlm(mpg ~ hp + drat, mtcars)
     pred <- predictions(model)
     expect_predictions(pred, n_row = nrow(mtcars))
@@ -159,7 +157,6 @@ test_that("glm.nb: marginalmeans: vs. emmeans", {
 
 
 test_that("rlm: marginalmeans: vs. emmeans", {
-    skip_if_not_installed("insight", minimum_version = "0.14.4.1")
     dat <- mtcars
     dat$cyl <- as.factor(dat$cyl)
     dat$am <- as.logical(dat$am)
@@ -174,13 +171,13 @@ test_that("rlm: marginalmeans: vs. emmeans", {
 
 
 test_that("polr: marginalmeans vs. emmeans", {
-    skip("works interactively")
     tmp <- mtcars
     tmp$vs <- as.factor(tmp$vs)
     tmp$am <- as.logical(tmp$am)
+    tmp <<- tmp
     mod <- suppressWarnings(MASS::polr(factor(gear) ~ vs + am, data = tmp))
     # TODO: emmeans seems broken at the moment
-    # em <- emmeans(mod, specs = "am", transform = "response")
+    # em <- emmeans(mod, specs = "am", regrid = "response")
     # em <- tidy(em)
     mm <- marginalmeans(mod, variables = "am", type = "probs")
     expect_equal(nrow(mm), 6)
@@ -212,10 +209,10 @@ test_that("glmmPQL: no validity", {
 
 
 test_that("bugs stay dead: character regressor with categorical outcome", {
-    skip("works interactively")
     requiet("MASS")
     dat <- mtcars
     dat$cyl <- as.character(dat$cyl)
+    dat <<- dat
     mod <- polr(factor(gear) ~ cyl, data = dat)
     # not clear why this generates a warning only on CI
     mfx <- suppressWarnings(marginaleffects(mod, type = "probs"))