+++ tests/testthat/test-plot.R
@@ -1,5 +1,4 @@
-skip_if(getRversion() < 4.1) # different graphics engines
-skip_on_ci()
+skip_if(getRversion() < "4.1.0") # different graphics engines
 skip_on_cran()
 
 test_that("plot_cme(mod, 'hp', 'wt')", {
@@ -13,6 +12,13 @@ test_that("plot(mfx)", {
     mfx <- marginaleffects(mod)
     p <- plot(mfx)
     vdiffr::expect_doppelganger("plot basic", p)
+})
+
+test_that("plot(mfx): no CI", {
+    mod <- glm(am ~ hp + wt, data = mtcars)
+    mfx <- marginaleffects(mod, vcov = FALSE)
+    p <- plot(mfx)
+    vdiffr::expect_doppelganger("plot no CI", p)
 })
 
 test_that("bugfix: contrasts overlap", {