+++ tests/testthat/test-plot_cap.R
@@ -1,4 +1,4 @@
-skip_on_ci() # different graphics engine produce different snapshots
+skip_if(getRversion() < "4.1.0") # different graphics engines
 skip_on_cran() # different graphics engine produce different snapshots
 
 test_that("two conditions", {
@@ -34,4 +34,11 @@ test_that("link vs response", {
 test_that("bad condition raises error", {
     mod <- lm(mpg ~ hp * wt * am, data = mtcars)
     expect_error(plot_cap(mod, condition = c("bad", "wt")))
+})
+
+test_that("Issue #230: glm w/ weights includes confidence intervals", {
+    mod <- glm(am ~ mpg * cyl, data = mtcars, family = binomial(link = "logit"), weights = carb)
+    p <- plot_cap(mod, condition = c("mpg", "cyl"), draw = FALSE)
+    expect_true("conf.low" %in% colnames(p))
+    expect_true("conf.high" %in% colnames(p))
 })